Moved all useful ElementDefinition extension methods to ElementDefinitionExtensions in the Model namespace.
